# docs/swagger.yaml
definitions:
...
  controllers.gameTagInput:
    properties:
      game_id:
        type: integer
      tag_id:
        type: integer
    type: object
  controllers.reviewInput:
    properties:
      description:
        type: string
      game_id:
        type: integer
      rating:
        type: integer
      user_id:
        type: integer
    type: object
  controllers.tagInput:
    properties:
      name:
        type: string
    type: object
...
  models.GameTag:
    properties:
      created_at:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      tag_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.Review:
    properties:
      created_at:
        type: string
      description:
        type: string
      game_id:
        type: integer
      id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Tag:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /game-tags:
    get:
      description: Get a list of GameTag.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GameTag'
            type: array
      summary: Get all gameTag.
      tags:
      - GameTag
    post:
      description: Creating a new GameTag.
      parameters:
      - description: the body to create a new GameTag
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.gameTagInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GameTag'
      security:
      - BearerToken: []
      summary: Create New GameTag.
      tags:
      - GameTag
  /game-tags/{id}:
    delete:
      description: Delete a gameTag by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: gameTag id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one gameTag.
      tags:
      - GameTag
    get:
      description: Get a GameTag by id.
      parameters:
      - description: gameTag id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GameTag'
      summary: Get GameTag.
      tags:
      - GameTag
...
  /reviews:
    get:
      description: Get a list of Review.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
      summary: Get all review.
      tags:
      - Review
    post:
      description: Creating a new Review.
      parameters:
      - description: the body to create a new Review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.reviewInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Create New Review.
      tags:
      - Review
  /reviews/{id}:
    delete:
      description: Delete a review by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Review.
      tags:
      - Review
    get:
      description: Get a Review by id.
      parameters:
      - description: review id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      summary: Get Review.
      tags:
      - Review
    patch:
      description: Update review by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: review id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an review
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.reviewInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
      security:
      - BearerToken: []
      summary: Update Review.
      tags:
      - Review
  /tags:
    get:
      description: Get a list of Tag.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tag'
            type: array
      summary: Get all Tag.
      tags:
      - Tag
    post:
      description: Creating a new Tag.
      parameters:
      - description: the body to create a new Tag
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.tagInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tag'
      security:
      - BearerToken: []
      summary: Create New Tag.
      tags:
      - Tag
  /tags/{id}:
    delete:
      description: Delete a Tag by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tag id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Tag.
      tags:
      - Tag
    get:
      description: Get an Tag by id.
      parameters:
      - description: Tag id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tag'
      summary: Get Tag.
      tags:
      - Tag
swagger: "2.0"